migrate test helpers to addon-test-support
